Fixes job insertion and retrieval inconsistencies
Addresses issues where job IDs were being handled as i64 instead of i32, causing potential type mismatch errors and incorrect data handling.

Changes the data type for job IDs to i32 throughout the backend to ensure consistency between database storage and application logic. This includes updating database queries and data structures to use i32 for job IDs.

Adds a debug log to the db initialization to help diagnose connection issues.

Also ensures standard output is flushed after printing the number of inserted and skipped jobs to ensure the logs are written to the console immediately.

+++ backend/src/handlers/jobs.rs
@@ -69,7 +69,7 @@ pub async fn create_job_handler(
         .map_err(|_| StatusCode::INTERNAL_SERVER_ERROR)?;
 
     let job = Job {
-        id: id as i32,
+        id: id, // Changed: removed "as i32"
         title: new_job.title,
         company: new_job.company,
         location: new_job.location,
@@ -125,9 +125,9 @@ pub async fn bulk_create_jobs_handler(
 
         // If RETURNING gives us an id, it was inserted
         if let Some(row) = result {
-            let id: i64 = row.get("id");
+            let id: i32 = row.get("id");
             let job = Job {
-                id: id as i32,
+                id: id,
                 title: new_job.title,
                 company: new_job.company,
                 location: new_job.location,
